+++ shell.nix
@@ -1,0 +1,29 @@
+let
+  moz_overlay = import (builtins.fetchTarball https://github.com/mozilla/nixpkgs-mozilla/archive/master.tar.gz);
+  nixpkgs = import <nixpkgs> { overlays = [ moz_overlay ]; };
+  rustStableChannel = nixpkgs.latest.rustChannels.stable.rust.override {
+    extensions = [
+      "rust-src"
+      "rls-preview"
+      "clippy-preview"
+      "rustfmt-preview"
+    ];
+  };
+in
+  with nixpkgs;
+  stdenv.mkDerivation {
+    name = "moz_overlay_shell";
+    nativeBuildInputs = [
+    pkgconfig
+    udev
+    dbus
+  ];
+  buildInputs = [
+    rustStableChannel
+    rls
+    openssl
+  ];
+
+  # Set Environment Variables
+  RUST_BACKTRACE = 1;
+}